requirements.txt
apache-airflow==2.9.3
psycopg2-binary
pandas
sqlalchemy
dbt-postgres
